requirements.txt

Django
djangorestframework          # api
drf-yasg                     # documentation